requirements.txt
# Web framework
fastapi==0.104.1
uvicorn==0.24.0

# Document processing
PyPDF2==3.0.1
python-multipart==0.0.6
aiofiles==23.2.1
python-docx==1.1.0
openpyxl==3.1.2
xlrd==2.0.1
Pillow==10.1.0
pytesseract==0.3.10

# Vector database and embeddings
chromadb==0.4.15
sentence-transformers==2.2.2

# LLM integration
openai==1.3.5
anthropic==0.7.7
groq==0.4.1
langchain==0.0.340
langchain-community==0.0.1

# Text processing
tiktoken==0.5.1
nltk==3.8.1

# Data handling
pandas==2.1.3
numpy==1.25.2

# Environment management
python-dotenv==1.0.0

# HTTP requests
httpx==0.25.2
requests==2.31.0

# Validation
pydantic==2.5.0

# Logging and monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Knowledge Graph & NLP
spacy==3.7.2
networkx==3.2.1
scikit-learn==1.3.2

# Document Art Generation
matplotlib==3.8.2
seaborn==0.13.0
wordcloud==1.9.2
plotly==5.17.0

# Additional utilities
jinja2==3.1.2
